requirements.txt
Django==4.2.25
requests==2.31.0
reportlab==4.0.7
Pillow==10.1.0
django-csp==3.7
beautifulsoup4>=4.11.0
lxml>=4.8.0
cryptography>=3.4.0

# Advanced Security Scanning Tools
python-nmap>=0.7.1
dnspython>=2.3.0
python-whois>=0.8.0
sslyze>=5.1.0
testssl>=3.0.8
sublist3r>=1.0
shodan>=1.28.0
censys>=2.1.0

# Web Application Security
selenium>=4.15.0
scrapy>=2.11.0
wappalyzer>=0.3.1
builtwith>=1.3.4

# Vulnerability Databases
cve-search-api>=1.0
nvdlib>=0.7.4
vulners>=2.0.9

# Network Analysis
scapy>=2.5.0
netaddr>=0.9.0
ipaddress>=1.0.23

# Advanced Reporting
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0
jinja2>=3.1.0

# Performance & Monitoring
psutil>=5.9.0
redis>=5.0.0
celery>=5.3.0

# Machine Learning for Threat Detection
scikit-learn>=1.3.0
pandas>=2.1.0
numpy>=1.24.0

# Additional Security Tools
hashlib-compat>=1.0.1
pyopenssl>=23.2.0
pycryptodome>=3.19.0

# Production Deployment
gunicorn>=21.2.0
whitenoise>=6.6.0
dj-database-url>=2.1.0
python-decouple>=3.8
psycopg2-binary>=2.9.7

# REST API
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
